+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-0.1.0:
+0.1.1:
 - Changed the core contracts to emit events while communicating with Zeus instead of using return types.
     - why: events are more highly structured, can be parsed via the ZeusScript ABI, and don't require manual tuple parsing.
 - `zeus env show <env>` now prints all of the contracts and environment parameters associated with an environment.